build(simengine-core): use correct *serv header based on OpenIPMI version
The change of migrating from serv.h to lanserv.h is in commit
c7b58da82e7176e556a33dee7104109bc2045e9f on the `code` repo of the
OpenIPMI project. Since the change was made after v2.0.30, we want to use
the "new" header (mcserv.h) when working with versions above v2.0.30.