fix(rpm): remove snmpsim patch
The reason for removing the patch for snmpsim is it doesn't actually solve
the problem. When SimEngine runs as root (uid=0), snmpsimd.py is always
set to run with user 'nobody' and group 'nobody'. The cause for the
exception (which was removed in the patch) only happens when user or group
is blank; but since 'nobody' is being passed to both, the exception should
never occur.